began fixes

+++ src/kern/drivers/keyboard.h
@@ -8,18 +8,18 @@
 
 struct keyboard
 {
-    uint64_t          keybuf[(4K >> 3)];
-    int32_t           keyvalbuf[(4K >> 3)];
+    long long          keybuf[(4K >> 3)];
+    long           keyvalbuf[(4K >> 3)];
     volatile long     keylk;
-    uint64_t         *curkey;
-    uint64_t         *lastkey;
-    int32_t          *curkeyval;
-    int32_t          *lastkeyval;
-    int32_t           keytab1b[128] __declspec(align((4K)));
-    int32_t           keytabmb[128];
-    int32_t           keytabup[128];
+    long long         *curkey;
+    long long         *lastkey;
+    long          *curkeyval;
+    long          *lastkeyval;
+    long           keytab1b[128] __declspec(align((4K)));
+    long           keytabmb[128];
+    long           keytabup[128];
     struct mousestate mousestate;
-    int32_t           modmask;
+    long           modmask;
 };
 
 inline void startKeyboard(void);
